Sprint 5 - OTC Tests

+++ otc/src/main/java/rs/edu/raf/si/bank2/otc/services/CompanyBankAccountService.java
@@ -5,9 +5,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rs.edu.raf.si.bank2.otc.dto.CompanyBankAccountDto;
 import rs.edu.raf.si.bank2.otc.exceptions.BankAccountNotFoundException;
+import rs.edu.raf.si.bank2.otc.exceptions.CompanyNotFoundException;
+import rs.edu.raf.si.bank2.otc.models.mongodb.Company;
 import rs.edu.raf.si.bank2.otc.models.mongodb.CompanyBankAccount;
 import rs.edu.raf.si.bank2.otc.repositories.mongodb.CompanyBankAccountRepository;
+import rs.edu.raf.si.bank2.otc.repositories.mongodb.CompanyRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,11 +21,14 @@ public class CompanyBankAccountService {
     @Autowired
     private CompanyBankAccountRepository bankAccountRepository;
 
+    @Autowired
+    private CompanyRepository companyRepository;
+
     public List<CompanyBankAccount> getAllBankAccounts() {
         return bankAccountRepository.findAll();
     }
 
-    public CompanyBankAccount getBankAccountById(Long id) {
+    public CompanyBankAccount getBankAccountById(String id) {
         Optional<CompanyBankAccount> bankAccount = bankAccountRepository.findById(id);
         if (bankAccount.isPresent()) {
             return bankAccount.get();
@@ -30,8 +37,30 @@ public class CompanyBankAccountService {
         }
     }
 
-    public CompanyBankAccount createBankAccount(CompanyBankAccount bankAccount) {
-        return bankAccountRepository.save(bankAccount);
+    public List<CompanyBankAccount> getAccountsForCompany(String companyId){
+        Optional<Company> companyOptional = companyRepository.findById(companyId);
+        if(companyOptional.isEmpty()) {
+            throw new CompanyNotFoundException("There is no company with this ID.");
+        }
+
+        return new ArrayList<>(companyOptional.get().getBankAccounts());
+    }
+
+    public CompanyBankAccount createBankAccount(String companyId, CompanyBankAccountDto accountDto) {
+        Optional<Company> companyOptional = companyRepository.findById(companyId);
+        if(companyOptional.isEmpty()) {
+            throw new CompanyNotFoundException("There is no company with this ID.");
+        }
+
+        CompanyBankAccount companyBankAccount = new CompanyBankAccount();
+            companyBankAccount.setAccountNumber(accountDto.getAccountNumber());
+            companyBankAccount.setBankName(accountDto.getBankName());
+            companyBankAccount.setCurrency(accountDto.getCurrency());
+        bankAccountRepository.save(companyBankAccount);
+        companyOptional.get().getBankAccounts().add(companyBankAccount);
+        companyRepository.save(companyOptional.get());
+
+        return companyBankAccount;
     }
 
     public CompanyBankAccount updateBankAccount(CompanyBankAccountDto bankAccountDto) {
@@ -40,18 +69,24 @@ public class CompanyBankAccountService {
             CompanyBankAccount companyBankAccount = existingBankAccount.get();
             companyBankAccount.setAccountNumber(bankAccountDto.getAccountNumber());
             companyBankAccount.setBankName(bankAccountDto.getBankName());
-            companyBankAccount.setAccountType(bankAccountDto.getAccountType());
             return bankAccountRepository.save(companyBankAccount);
         } else {
             throw new BankAccountNotFoundException("Bank account not found with id: " + bankAccountDto.getId());
         }
     }
 
-    public void deleteBankAccount(Long id) {
+    public void deleteBankAccount(String id, String companyId) {
+        Optional<Company> companyOptional = companyRepository.findById(companyId);
+        if(companyOptional.isEmpty()) {
+            throw new CompanyNotFoundException("There is no company with this ID.");
+        }
         Optional<CompanyBankAccount> bankAccount = bankAccountRepository.findById(id);
         if (bankAccount.isPresent()) {
+            companyOptional.get().getBankAccounts().remove(bankAccount.get());
+            companyRepository.save(companyOptional.get());
             bankAccountRepository.deleteById(id);
-        } else {
+        }
+        else {
             throw new BankAccountNotFoundException("Bank account not found with id: " + id);
         }
     }
